+++ src/nsidc/metgen/metgen.py
@@ -136,6 +136,14 @@ def init_config(configuration_file):
             default=str(constants.DEFAULT_COLLECTION_GEOMETRY_OVERRIDE),
         ),
     )
+    cfg_parser.set(
+        constants.SOURCE_SECTION_NAME,
+        "collection_temporal_override",
+        Prompt.ask(
+            "Use collection temporal extent? (True/False)",
+            default=str(constants.DEFAULT_COLLECTION_TEMPORAL_OVERRIDE),
+        ),
+    )
     print()
 
     print()
@@ -231,10 +239,6 @@ def init_config(configuration_file):
 class Collection:
     """
     Collection metadata relevant to generating UMM-G content
-
-    spatial_extent and temporal_extent currently contain the relevant
-    JSON structure from the collection metadata (if it exists).
-    Additional work is still needed to parse out the relevant fields.
     """
 
     auth_id: str
@@ -487,7 +491,7 @@ def validate_cmr_response(umm: list):
     return ummc
 
 
-def ummc_content(ummc: dict, keys: list) -> str | dict:
+def ummc_content(ummc: dict, keys: list) -> str | list | dict:
     """
     Look for list of keys in a UMM-C record and log the status.
     """
@@ -562,8 +566,27 @@ def collection_from_cmr(environment: str, auth_id: str, version: int):
             ummc, constants.GRANULE_SPATIAL_REP_PATH
         ),
         spatial_extent=ummc_content(ummc, constants.SPATIAL_EXTENT_PATH),
-        temporal_extent=ummc_content(ummc, constants.TEMPORAL_EXTENT_PATH),
+        temporal_extent=ummc_temporal_content(ummc),
     )
+
+
+def ummc_temporal_content(ummc: dict) -> str | dict:
+    # TemporalExtents points to an array, not a dict
+    temp = ummc_content(ummc, constants.TEMPORAL_EXTENT_PATH)
+    if len(temp) > 1:
+        raise Exception("too many temporal")
+
+    # single value will be represented as a list containing one string
+    temporal = ummc_content(temp[0], constants.TEMPORAL_SINGLE_PATH)
+
+    if temporal is None:
+        # range represented as a list containing one dict, dict has
+        # both begin and end keys even if end value is None
+        temporal = ummc_content(temp[0], constants.TEMPORAL_RANGE_PATH)
+
+    print(f"temporal is {temporal}")
+
+    return temporal
 
 
 def grouped_granule_files(configuration: config.Config) -> list[tuple]:
@@ -719,11 +742,34 @@ def validate_collection(configuration: config.Config, granule: Granule) -> Granu
     """
     Confirm collection metadata meet requirements for our granule metadata generation.
     """
-    errors = validate_collection_spatial(configuration, granule.collection)
-    if len(errors) == 0:
-        return granule
-    else:
+    errors = validate_collection_spatial(
+        configuration, granule.collection
+    ) + validate_collection_temporal(configuration, granule.collection)
+    if errors:
         raise config.ValidationError(errors)
+
+    return granule
+
+
+def validate_collection_temporal(configuration, collection):
+    """
+    Verify collection temporal extent information can be used if a collection
+    override is requested.
+    """
+    errors = []
+
+    # If collection temporal extent is to be applied to granules, there must be only
+    # one temporal extent object, which can be either a range or a single time value.
+    # A temporal range must include a BeginningDateTime, and may include an EndingDateTime.
+    if (
+        configuration.collection_temporal_override
+        and len(collection.temporal_extent) > 1
+    ):
+        errors.append(
+            "Collection metadata must only contain one temporal extent when collection_temporal_override is set."
+        )
+
+    return errors
 
 
 def validate_collection_spatial(configuration, collection):
@@ -740,7 +786,7 @@ def validate_collection_spatial(configuration, collection):
         )
 
     # If collection spatial extent is to be applied to granules, the spatial extent may
-    # only contain one bounding rectange, and the granule spatial representation must be cartesian
+    # only contain one bounding rectangle, and the granule spatial representation must be cartesian
     if configuration.collection_geometry_override:
         if not collection.spatial_extent:
             errors.append("Collection must include a spatial extent.")
@@ -800,6 +846,9 @@ def create_ummg(configuration: config.Config, granule: Granule) -> Granule:
 
     # Get premet content if it exists.
     premet_content = utilities.premet_values(granule.premet_filename)
+    temporal_content = utilities.external_temporal_values(
+        configuration.collection_temporal_override, premet_content, granule
+    )
 
     # Get spatial coverage from spatial file if it exists
     spatial_content = utilities.external_spatial_values(
@@ -825,7 +874,7 @@ def create_ummg(configuration: config.Config, granule: Granule) -> Granule:
             ),
         } | granule.data_reader(
             data_file,
-            premet_content,
+            temporal_content,
             spatial_content,
             configuration,
             gsr,
@@ -1095,9 +1144,10 @@ def populate_temporal(datetime_values):
     """
     Return a string representation of a temporal range or single value, as appropriate.
     """
-    if len(datetime_values) > 1:
+    if isinstance(datetime_values[0], dict):
+        # range_parts = [': '.join([k, str(v)]) for k, v in datetime_values[0].items()]
         return ummg_temporal_range_template().safe_substitute(
-            {"begin_date_time": datetime_values[0], "end_date_time": datetime_values[1]}
+            {"date_time_range": json.dumps(datetime_values[0])}
         )
     else:
         return ummg_temporal_single_template().safe_substitute(