+++ src/nsidc/metgen/readers/netcdf_reader.py
@@ -20,7 +20,7 @@ from nsidc.metgen.readers import utilities
 
 def extract_metadata(
     netcdf_path: str,
-    premet_content: dict,
+    temporal_content: list,
     spatial_content: list,
     configuration: Config,
     gsr: str,
@@ -48,8 +48,8 @@ def extract_metadata(
         raise Exception(f"Could not open netCDF file {netcdf_path}")
 
     # Use temporal coverage from premet file if it exists
-    if premet_content:
-        temporal = utilities.temporal_from_premet(premet_content)
+    if temporal_content:
+        temporal = temporal_content
     else:
         temporal = time_range(os.path.basename(netcdf_path), netcdf, configuration)
 
@@ -73,7 +73,7 @@ def time_range(netcdf_filename, netcdf, configuration):
     coverage_end = time_coverage_end(netcdf, configuration, coverage_start)
 
     if coverage_start and coverage_end:
-        return [coverage_start, coverage_end]
+        return utilities.refine_temporal([coverage_start, coverage_end])
     else:
         # In theory, we should never get here.
         log_and_raise_error(